[refactor/#31] 상품 등록 API 통합 및 중복 파일 정리
- registerProduct 함수 productApi.ts 내로 통합
- 불필요한 중복 API 파일 제거
- 공통 API 관리 구조 개선
